Provide direct link to storage providers
Add a link directly to the storage providers page from  the introduction
section on supported backends.
Remove OneDrive and SharePoint v2 from the supported backends list
because these are not real backends, they are just referring to the
different versions of the backend, not actual different product versions
Details about the backend are avaliable on the Storage Providers page.