Update installation instructions to cater also for Debian 10 buster.
Have updated the installation instructions to cater also for Debian 10 buster. Note that the Mono installation instructions advice buster-stable rather than just buster. But that is true for earlier versions also.

Note also that the instructions for downloading and installing the Debian package setting it up as a service that auto-starts also need to be added. I'll try to add those next.